some refactoring, sound working

+++ src/subslim/synth/effect/AllPassFilter.java
@@ -16,7 +16,7 @@ public class AllPassFilter {
 
         delayBuffer = new LinkedList();
 
-        for(int i = 0; i < (delay/1000)* Synth.SAMPLE_RATE; i++){
+        for(int i = 0; i < (delay/1000d)* (double) Synth.SAMPLE_RATE; i++){
             delayBuffer.addFirst(0d);
         }
